test: move tests for AddCb to separate files
